+++ NEWS.md
@@ -3,4 +3,4 @@
 NEW FEATURES
 
 * Added a `NEWS.md` file to track changes to the package.
-* Deprecated `run_ewce()`, on account of major changes to the structure of EWCE, which made this function redundant.
+* Deprecated `run_ewce()`, on account of major changes to the structure of EWCE, which made this function redundant. See [MultiEWCE](https://github.com/neurogenomics/MultiEWCE/) and EWCE [news](https://github.com/NathanSkene/EWCE/blob/master/NEWS.md).